% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipelineBuilder.R
\name{dataPipelineBuilder}
\alias{dataPipelineBuilder}
\title{Data Pipeline Builder function}
\usage{
dataPipelineBuilder()
}
\description{
`dataPipelineBuilder` builds and processes a pipeline of data processing nodes.
}
\details{
TODO
}
