% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipelineNode.R
\name{DataPipelineNode}
\alias{DataPipelineNode}
\title{Data Pipeline Node R6Class}
\description{
`DataPipelineNode` is an R6Class which support the generic methods for managing a pipeline.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataPipelineNode-new}{\code{DataPipelineNode$new()}}
\item \href{#method-DataPipelineNode-print}{\code{DataPipelineNode$print()}}
\item \href{#method-DataPipelineNode-processData}{\code{DataPipelineNode$processData()}}
\item \href{#method-DataPipelineNode-clone}{\code{DataPipelineNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataPipelineNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataPipelineNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataPipelineNode$new(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataPipelineNode-print"></a>}}
\if{latex}{\out{\hypertarget{method-DataPipelineNode-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataPipelineNode$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataPipelineNode-processData"></a>}}
\if{latex}{\out{\hypertarget{method-DataPipelineNode-processData}{}}}
\subsection{Method \code{processData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataPipelineNode$processData(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataPipelineNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataPipelineNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataPipelineNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
