% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineBuilder.R
\name{.loadAndValidateConfig}
\alias{.loadAndValidateConfig}
\title{Data Pipeline Builder function}
\usage{
.loadAndValidateConfig(configFilename)
}
\description{
`dataPipelineBuilder` builds and processes a pipeline of data processing nodes.
}
\details{
TODO
}
