% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromLongLongDataPipelineNode.R
\name{FromLongLongDataPipelineNode}
\alias{FromLongLongDataPipelineNode}
\title{From 'long long' Data Pipeline Node R6Class}
\description{
`FromLongLongDataPipelineNode` is an R6Class which support the conversion from 'long long' format of data which passes to the `processData` function by the previous node in the pipeline.
}
\details{
TODO
}
\section{Super class}{
\code{\link[DataProcessingPipeline:DataPipelineNode]{DataProcessingPipeline::DataPipelineNode}} -> \code{FromLongLongDataPipelineNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FromLongLongDataPipelineNode-new}{\code{FromLongLongDataPipelineNode$new()}}
\item \href{#method-FromLongLongDataPipelineNode-processData}{\code{FromLongLongDataPipelineNode$processData()}}
\item \href{#method-FromLongLongDataPipelineNode-clone}{\code{FromLongLongDataPipelineNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DataProcessingPipeline" data-topic="DataPipelineNode" data-id="print"><a href='../../DataProcessingPipeline/html/DataPipelineNode.html#method-DataPipelineNode-print'><code>DataProcessingPipeline::DataPipelineNode$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromLongLongDataPipelineNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-FromLongLongDataPipelineNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromLongLongDataPipelineNode$new(name, config)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromLongLongDataPipelineNode-processData"></a>}}
\if{latex}{\out{\hypertarget{method-FromLongLongDataPipelineNode-processData}{}}}
\subsection{Method \code{processData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromLongLongDataPipelineNode$processData(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FromLongLongDataPipelineNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FromLongLongDataPipelineNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FromLongLongDataPipelineNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
