% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoaderDataPipelineNode.R
\name{LoaderDataPipelineNode}
\alias{LoaderDataPipelineNode}
\title{Loader Data Pipeline Node R6Class}
\description{
`LoaderDataPipelineNode` is an R6Class which support the loading of data which passes to the `processData` function of the next node in the pipeline.
}
\details{
TODO
}
\section{Super class}{
\code{\link[DataProcessingPipeline:DataPipelineNode]{DataProcessingPipeline::DataPipelineNode}} -> \code{LoaderDataPipelineNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LoaderDataPipelineNode-new}{\code{LoaderDataPipelineNode$new()}}
\item \href{#method-LoaderDataPipelineNode-processData}{\code{LoaderDataPipelineNode$processData()}}
\item \href{#method-LoaderDataPipelineNode-clone}{\code{LoaderDataPipelineNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DataProcessingPipeline" data-topic="DataPipelineNode" data-id="print"><a href='../../DataProcessingPipeline/html/DataPipelineNode.html#method-DataPipelineNode-print'><code>DataProcessingPipeline::DataPipelineNode$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoaderDataPipelineNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-LoaderDataPipelineNode-new}{}}}
\subsection{Method \code{new()}}{
Initialize the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoaderDataPipelineNode$new(name, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of node}

\item{\code{config}}{config of node}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoaderDataPipelineNode-processData"></a>}}
\if{latex}{\out{\hypertarget{method-LoaderDataPipelineNode-processData}{}}}
\subsection{Method \code{processData()}}{
Process data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoaderDataPipelineNode$processData(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data to be processed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoaderDataPipelineNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LoaderDataPipelineNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoaderDataPipelineNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
