% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryDataPipelineNode.R
\name{SummaryDataPipelineNode}
\alias{SummaryDataPipelineNode}
\title{Summary Data Pipeline Node R6Class}
\description{
`SummaryDataPipelineNode` is an R6Class which support the summarising ata which passes to the `processData` function by the previous node in the pipeline.
}
\details{
TODO
}
\section{Super class}{
\code{\link[DataProcessingPipeline:DataPipelineNode]{DataProcessingPipeline::DataPipelineNode}} -> \code{SummaryDataPipelineNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SummaryDataPipelineNode-new}{\code{SummaryDataPipelineNode$new()}}
\item \href{#method-SummaryDataPipelineNode-processData}{\code{SummaryDataPipelineNode$processData()}}
\item \href{#method-SummaryDataPipelineNode-clone}{\code{SummaryDataPipelineNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DataProcessingPipeline" data-topic="DataPipelineNode" data-id="print"><a href='../../DataProcessingPipeline/html/DataPipelineNode.html#method-DataPipelineNode-print'><code>DataProcessingPipeline::DataPipelineNode$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SummaryDataPipelineNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-SummaryDataPipelineNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SummaryDataPipelineNode$new(name, config)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SummaryDataPipelineNode-processData"></a>}}
\if{latex}{\out{\hypertarget{method-SummaryDataPipelineNode-processData}{}}}
\subsection{Method \code{processData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SummaryDataPipelineNode$processData(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SummaryDataPipelineNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SummaryDataPipelineNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SummaryDataPipelineNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
