% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidaterDataPipelineNode.R
\name{ValidaterDataPipelineNode}
\alias{ValidaterDataPipelineNode}
\title{Validater Data Pipeline Node R6Class}
\description{
`ValidaterDataPipelineNode` is an R6Class which support the validating of data which passes to the `processData` function by the previous node in the pipeline.
}
\details{
TODO
}
\section{Super class}{
\code{\link[DataProcessingPipeline:DataPipelineNode]{DataProcessingPipeline::DataPipelineNode}} -> \code{ValidaterDataPipelineNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValidaterDataPipelineNode-new}{\code{ValidaterDataPipelineNode$new()}}
\item \href{#method-ValidaterDataPipelineNode-processData}{\code{ValidaterDataPipelineNode$processData()}}
\item \href{#method-ValidaterDataPipelineNode-clone}{\code{ValidaterDataPipelineNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DataProcessingPipeline" data-topic="DataPipelineNode" data-id="print"><a href='../../DataProcessingPipeline/html/DataPipelineNode.html#method-DataPipelineNode-print'><code>DataProcessingPipeline::DataPipelineNode$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidaterDataPipelineNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-ValidaterDataPipelineNode-new}{}}}
\subsection{Method \code{new()}}{
Initialize the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidaterDataPipelineNode$new(name, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of node}

\item{\code{config}}{config of node}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidaterDataPipelineNode-processData"></a>}}
\if{latex}{\out{\hypertarget{method-ValidaterDataPipelineNode-processData}{}}}
\subsection{Method \code{processData()}}{
Process data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidaterDataPipelineNode$processData(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data to be processed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidaterDataPipelineNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValidaterDataPipelineNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValidaterDataPipelineNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
