% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteTableDataPipelineNode.R
\name{WriteTableDataPipelineNode}
\alias{WriteTableDataPipelineNode}
\title{WriteTable Data Pipeline Node R6Class}
\description{
`WriteTableDataPipelineNode` is an R6Class which support the writing of table data which passes to the `processData` function of the next node in the pipeline.
}
\details{
TODO
}
\section{Super class}{
\code{\link[DataProcessingPipeline:DataPipelineNode]{DataProcessingPipeline::DataPipelineNode}} -> \code{WriteTableDataPipelineNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WriteTableDataPipelineNode-new}{\code{WriteTableDataPipelineNode$new()}}
\item \href{#method-WriteTableDataPipelineNode-processData}{\code{WriteTableDataPipelineNode$processData()}}
\item \href{#method-WriteTableDataPipelineNode-clone}{\code{WriteTableDataPipelineNode$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DataProcessingPipeline" data-topic="DataPipelineNode" data-id="print"><a href='../../DataProcessingPipeline/html/DataPipelineNode.html#method-DataPipelineNode-print'><code>DataProcessingPipeline::DataPipelineNode$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WriteTableDataPipelineNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-WriteTableDataPipelineNode-new}{}}}
\subsection{Method \code{new()}}{
Initialize the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WriteTableDataPipelineNode$new(name, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of node}

\item{\code{config}}{config of node}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WriteTableDataPipelineNode-processData"></a>}}
\if{latex}{\out{\hypertarget{method-WriteTableDataPipelineNode-processData}{}}}
\subsection{Method \code{processData()}}{
Process data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WriteTableDataPipelineNode$processData(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data to be processed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WriteTableDataPipelineNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WriteTableDataPipelineNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WriteTableDataPipelineNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
